Data Wrangling Exercise 1 script
This is my script for the first Data Wrangling Exercise! I had a weird issue with unilver not being changed to unilever while the others did,
and where Laptop in the product category column would become LaSmartphoneptoSMartphonetop depending on the order I ran those 4 gsub lines,
and I didn't exactly understand the dummy part until I did some googling and found something on rbloggers which used a for loop. Woo